move FamilyBuilder into Family.Builder
