+++ src/main/java/gedcom/models/Family.java
@@ -1,6 +1,8 @@
 package gedcom.models;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Calendar;
 import java.util.Comparator;
 import java.util.Date;
 import java.util.HashSet;
@@ -141,6 +143,165 @@ public class Family extends GEDObject {
 
     public boolean isChild(Individual individual) {
         return children.contains(individual);
+    }
+
+    public static Builder builder() {
+        return new Builder();
+    }
+
+    public static final class Builder extends GEDObjectBuilder {
+
+        private static int id = 1;
+
+        private String ID;
+
+        private Individual husband;
+        private Individual wife;
+
+        private Date marriage;
+        private Date divorce;
+
+        private List<Individual> children;
+
+        // A record of the families this builder has created
+        private List<Family> families;
+
+        private Builder() {
+            this.children = new ArrayList<>();
+            this.families = new ArrayList<>();
+        }
+
+        public Builder ID(String ID) {
+            this.ID = ID;
+            return this;
+        }
+
+        public Builder husband(Individual husband) {
+            this.husband = husband;
+            return this;
+        }
+
+        public Builder wife(Individual wife) {
+            this.wife = wife;
+            return this;
+        }
+
+        public Builder spouses(Individual husband, Individual wife) {
+            this.husband = husband;
+            this.wife = wife;
+            return this;
+        }
+
+        public Builder child(Individual child) {
+            this.children.add(child);
+            return this;
+        }
+
+        public Builder children(List<Individual> children) {
+            this.children.addAll(children);
+            return this;
+        }
+
+        public Builder children(Individual... children) {
+            this.children.addAll(Arrays.asList(children));
+            return this;
+        }
+
+        public Builder children(int n) {
+            for (int i = 0; i < n; i++) {
+                this.children.add(Individual.builder().build());
+            }
+            return this;
+        }
+
+        public Builder marriage(Date marriage) {
+            this.marriage = marriage;
+            return this;
+        }
+
+        public Builder marriage(int date, int month, int year) {
+            this.marriage = buildDate(date, month, year);
+            return this;
+        }
+
+        public Builder marriage(int year) {
+            this.marriage = buildDate(1, Calendar.JANUARY, year);
+            return this;
+        }
+
+        public Builder divorce(Date divorce) {
+            this.divorce = divorce;
+            return this;
+        }
+
+        public Builder divorce(int date, int month, int year) {
+            this.divorce = buildDate(date, month, year);
+            return this;
+        }
+
+        public Builder divorce(int year) {
+            this.divorce = buildDate(1, Calendar.JANUARY, year);
+            return this;
+        }
+
+        public Family build() {
+
+            Family family;
+
+            if (ID != null) {
+                family = new Family(ID);
+            } else {
+                family = new Family(String.format("@F%d@", id++));
+            }
+
+            if (husband != null) {
+                family.setHusband(husband);
+                husband.addSpouseFamily(family);
+            }
+
+            if (wife != null) {
+                family.setWife(wife);
+                wife.addSpouseFamily(family);
+            }
+
+            for (Individual child : children) {
+                child.addChildFamily(family);
+                family.addChild(child);
+            }
+
+            if (marriage != null) {
+                family.setMarriage(marriage);
+            }
+
+            if (divorce != null) {
+                family.setDivorce(divorce);
+            }
+
+            this.families.add(family);
+
+            clear();
+
+            return family;
+        }
+
+        private void clear() {
+            this.ID = null;
+            this.husband = null;
+            this.wife = null;
+            this.marriage = null;
+            this.divorce = null;
+            this.children = new ArrayList<>();
+        }
+
+        /**
+         * Returns the list of all Families built by this builder.
+         * 
+         * @return
+         */
+        public List<Family> getFamilies() {
+            return this.families;
+        }
+
     }
 
 }
